refactor(services): mv fnIngress to indexer-next and remove npm package
