--- addons/services/lambdas/fnIngress/package.json
@@ -1,13 +1,0 @@
-{
-	"type": "module",
-	"dependencies": {
-		"@aws-sdk/client-s3": "^3.1027.0",
-		"@aws-sdk/lib-storage": "^3.1027.0",
-		"@aws-sdk/node-http-handler": "^3.370.0",
-		"ar-gql": "^3.1.0",
-		"file-type": "^21.0.0",
-		"moize": "^6.1.6",
-		"pg": "^8.13.0",
-		"shepherd-plugin-interfaces": "^2.0.0"
-	}
-}
