--- addons/services/lambdas/fnIngress/tsconfig.json
@@ -1,10 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES2022",
-		"module": "ES2022",
-		"moduleResolution": "bundler",
-		"strict": true,
-		"esModuleInterop": true,
-		"skipLibCheck": true,
-	}
-}
